pyproject.toml
[tool.poetry]
name = "pythonproject9"
version = "0.1.0"
description = ""
authors = ["Arina Astanina <timyr.kiyan@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.9"
python-dotenv = "^1.0.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

